requirements.txt
Django
gunicorn
django-allauth
python-dotenv
whitenoise
django-bootstrap-v5
django-bootstrap-icons
postgres
googlemaps